backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
groq>=0.11.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
